Update gengokit to have config in base with generator package
