--- gengokit/config/config.go
@@ -1,10 +1,0 @@
-package config
-
-import "github.com/TuneLab/go-truss/truss/truss"
-
-type Config struct {
-	GoPackage string
-	PBPackage string
-
-	PreviousFiles []truss.NamedReadWriter
-}
