+++ truss/main.go
@@ -17,8 +17,8 @@ import (
 
 	"github.com/TuneLab/go-truss/deftree"
 	"github.com/TuneLab/go-truss/gendoc"
-	"github.com/TuneLab/go-truss/gengokit"
-	ggkconf "github.com/TuneLab/go-truss/gengokit/config"
+	ggkconf "github.com/TuneLab/go-truss/gengokit"
+	gengokit "github.com/TuneLab/go-truss/gengokit/generator"
 	"github.com/TuneLab/go-truss/svcdef"
 )
